Add files via upload
This here is to download the "Motor Vehicle Collisions involving Killed or Seriously Injured Persons" file we found. Some changes will have to be made (for example, the df is not yet a geopandas df), but I think its good for a start and also gets us started with GitHub as well.